+++ ingestion_tools/scripts/data_validation/fixtures/path.py
@@ -83,9 +83,9 @@ def frames_dir(run_dir: str, filesystem: FileSystemApi) -> str:
 def frames_files(frames_dir: str, filesystem: FileSystemApi) -> List[str]:
     """[Dataset]/[ExperimentRun]/Frames/*"""
     files = filesystem.glob(f"{frames_dir}/*")
-    # Exclude gain files, add s3 prefix
-    refined_files = ["s3://" + file for file in files if "_gain" not in file]
-    # gain files are in the folder, but just no frames
+    # Exclude mdoc files, add s3 prefix
+    refined_files = ["s3://" + file for file in files if ".mdoc" not in file]
+    # mdoc files are in the folder, but just no frames
     if len(refined_files) == 0 and len(files) != 0:
         pytest.skip(f"No frame files in frames directory: {frames_dir}")
 
@@ -100,7 +100,7 @@ def frames_files(frames_dir: str, filesystem: FileSystemApi) -> List[str]:
 @pytest.fixture(scope="session")
 def gain_dir(run_dir: str, filesystem: FileSystemApi) -> str:
     """[Dataset]/[ExperimentRun]/Frames"""
-    dst = f"{run_dir}/Frames"
+    dst = f"{run_dir}/Gains"
     if filesystem.exists(dst):
         return dst
     else:
@@ -110,12 +110,11 @@ def gain_dir(run_dir: str, filesystem: FileSystemApi) -> str:
 @pytest.fixture(scope="session")
 def gain_files(run_name: str, gain_dir: str, filesystem: FileSystemApi) -> List[str]:
     """[Dataset]/[ExperimentRun]/Frames/[run_name]*_gain*"""
-    files = filesystem.glob(f"{gain_dir}/*_gain*")
+    files = filesystem.glob(f"{gain_dir}/*")
     if len(files) == 0:
         pytest.skip("No gain files found.")
 
     for file in files:
-        assert f"{run_name}_gain" in file, f"Invalid gain file name: {file}"
         assert not file.endswith(".dm4"), f"Invalid gain file extension: {file} (should be .mrc)"
 
     return files
