+++ ingestion_tools/scripts/data_validation/tests/test_tilt_angles.py
@@ -1,10 +1,12 @@
 import os
-from typing import Dict, List
+from typing import Dict, List, Union
 
 import allure
 import numpy as np
 import pandas as pd
 import pytest
+import tifffile
+from mrcfile.mrcinterpreter import MrcInterpreter
 
 ANGLE_TOLERANCE = 0.05
 
@@ -109,10 +111,10 @@ class TestTiltAngles:
         assert len(errors) == 0, "\n".join(errors)
 
     @allure.title("Tilt: every tilt angle maps to a mdoc tilt angle.")
-    def test_tilt_mdoc(self, tiltseries_tilt: pd.DataFrame, tiltseries_mdoc: pd.DataFrame):
+    def test_tilt_mdoc(self, tiltseries_tilt: pd.DataFrame, frames_mdoc: pd.DataFrame):
         errors = self.helper_angles_injection_errors(
             tiltseries_tilt["TiltAngle"].to_list(),
-            tiltseries_mdoc["TiltAngle"].to_list(),
+            frames_mdoc["TiltAngle"].to_list(),
             "tilt file",
             "mdoc file",
         )
@@ -150,10 +152,10 @@ class TestTiltAngles:
         assert all(-90 <= angle <= 90 for angle in tiltseries_raw_tilt["TiltAngle"])
 
     @allure.title("Raw tilt: every raw tilt angle matches a mdoc tilt angle.")
-    def test_raw_tilt_mdoc(self, tiltseries_raw_tilt: pd.DataFrame, tiltseries_mdoc: pd.DataFrame):
+    def test_raw_tilt_mdoc(self, tiltseries_raw_tilt: pd.DataFrame, frames_mdoc: pd.DataFrame):
         errors = self.helper_angles_injection_errors(
             tiltseries_raw_tilt["TiltAngle"].to_list(),
-            tiltseries_mdoc["TiltAngle"].to_list(),
+            frames_mdoc["TiltAngle"].to_list(),
             "raw tilt file",
             "mdoc file",
         )
@@ -182,56 +184,18 @@ class TestTiltAngles:
         )
 
     ### BEGIN MDOC tests ###
-    @allure.title("Mdoc: tilt angles are within the expected range [-90, 90].")
-    def test_tiltseries_mdoc_range(self, tiltseries_mdoc: pd.DataFrame):
-        assert tiltseries_mdoc["TiltAngle"].min() >= -90
-        assert tiltseries_mdoc["TiltAngle"].max() <= 90
-
-    @allure.title("Mdoc: number of mdoc tilt angles equals number of frames files.")
-    def test_mdoc_frames(self, tiltseries_mdoc: pd.DataFrame, frames_files: List[str]):
-        assert len(tiltseries_mdoc) == len(frames_files)
 
     @allure.title("Mdoc: number of mdoc tilt angles equals tiltseries size['z'].")
-    def test_mdoc_tiltseries_metadata(self, tiltseries_metadata: Dict, tiltseries_mdoc: pd.DataFrame):
-        assert len(tiltseries_mdoc) == tiltseries_metadata["size"]["z"]
-
-    @allure.title("Mdoc: Every mdoc tilt angle matches a frames file (one-to-one).")
-    def test_mdoc_frame_paths(
-        self,
-        frames_files: List[str],
-        tiltseries_mdoc: pd.DataFrame,
-    ):
-        standardize_frames_files = [os.path.basename(f) for f in frames_files]
-        standardized_mdoc_entries = [
-            os.path.basename(str(row["SubFramePath"]).replace("\\", "/")) for _, row in tiltseries_mdoc.iterrows()
-        ]
-
-        mdoc_with_missing_frames = [
-            entry for entry in standardized_mdoc_entries if entry not in standardize_frames_files
-        ]
-        frames_with_missing_mdoc = [
-            entry for entry in standardize_frames_files if entry not in standardized_mdoc_entries
-        ]
-
-        errors = []
-        if len(standardize_frames_files) != len(standardized_mdoc_entries):
-            errors.append(
-                f"# of mdoc entries ({len(standardized_mdoc_entries)}) != # of frames files ({len(standardize_frames_files)})",
-            )
-        if mdoc_with_missing_frames:
-            errors.append(f"mdoc entries do not have frames files: {mdoc_with_missing_frames}")
-        if frames_with_missing_mdoc:
-            errors.append(f"Frames files do not have mdoc entries: {frames_with_missing_mdoc}")
-
-        assert len(errors) == 0, "\n".join(errors)
+    def test_mdoc_tiltseries_metadata(self, tiltseries_metadata: Dict, frames_mdoc: pd.DataFrame):
+        assert len(frames_mdoc) == tiltseries_metadata["size"]["z"]
 
     @allure.title("Mdoc: every mdoc tilt angle corresponds to the tilt_range + tilt_step metadata field.")
     @allure.description("Not all angles in the tilt range must be present in the MDOC file.")
     def test_mdoc_tiltseries_range(
-        self, tiltseries_metadata: Dict, tiltseries_mdoc: pd.DataFrame, tiltseries_metadata_range: List[float],
+        self, tiltseries_metadata: Dict, frames_mdoc: pd.DataFrame, tiltseries_metadata_range: List[float],
     ):
         errors = self.helper_angles_injection_errors(
-            tiltseries_mdoc["TiltAngle"].to_list(),
+            frames_mdoc["TiltAngle"].to_list(),
             tiltseries_metadata_range,
             "mdoc file",
             "tiltseries metadata tilt_range",
@@ -240,3 +204,22 @@ class TestTiltAngles:
             "\n".join(errors)
             + f"\nRange: {tiltseries_metadata['tilt_range']['min']} to {tiltseries_metadata['tilt_range']['max']}, with step {tiltseries_metadata['tilt_step']}"
         )
+
+    ### BEGIN Tiltseries consistency tests ###
+    @allure.title("Frames: tiltseries pixel spacing is an integer multiple of the frame pixel spacing.")
+    def test_tiltseries_pixel_spacing(
+        self,
+        dataset: str,
+        frames_headers: Dict[str, Union[List[tifffile.TiffPage], MrcInterpreter]],
+        tiltseries_metadata: Dict,
+    ):
+        for frame_file, frame_header in frames_headers.items():
+            if isinstance(frame_header, MrcInterpreter):
+                # only need to check the first frame, since we check that all frames have the same pixel spacing
+                assert tiltseries_metadata["pixel_spacing"] / frame_header.voxel_size["x"] == pytest.approx(
+                    round(tiltseries_metadata["pixel_spacing"] / frame_header.voxel_size["x"]),
+                    abs=0.001,
+                ), f"Pixel spacing does not match tiltseries metadata, {frame_file}"
+                return
+
+    ### END Tiltseries consistency tests ###
